feat: add version information flag
Add a flag that will print the version of Woodwork when running the CLI.
It pulls this version from the pyproject.toml file.
It will end the program run after printing the version.
