+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "pyyaml",
     "jsonpointer",
     "deprecated",
+    "tomli>=2.2.1",
 ]
 classifiers = [
     "Programming Language :: Python :: 3.10",
